Add initial testing setup
