+++ apps/backend/jest.config.js
@@ -1,0 +1,5 @@
+/** @type {import('ts-jest').InitialOptionsTsJest} */
+module.exports = {
+  preset: "ts-jest",
+  testEnvironment: "node",
+};
