+++ apps/backend/package.json
@@ -6,7 +6,8 @@
     "scripts": {
         "dev": "nodemon src/index.ts",
         "build": "tsc",
-        "lint": "eslint src --ext .js,.jsx,.ts,.tsx"
+        "lint": "eslint src --ext .js,.jsx,.ts,.tsx",
+        "test": "jest"
     },
     "dependencies": {
         "@prisma/client": "^3.11.0",
@@ -21,11 +22,16 @@
         "@types/compression": "^1.7.2",
         "@types/cookie-parser": "^1.4.2",
         "@types/express": "^4.17.13",
+        "@types/jest": "^27.4.1",
         "@types/node": "^17.0.21",
+        "@types/supertest": "^2.0.11",
         "config": "*",
         "eslint": "^8.10.0",
+        "jest": "^27.5.1",
         "nodemon": "^2.0.15",
         "prisma": "^3.11.0",
+        "supertest": "^6.2.2",
+        "ts-jest": "^27.1.3",
         "ts-node": "^10.5.0",
         "tsconfig": "*",
         "typescript": "^4.5.5"
