+++ apps/frontend/package.json
@@ -28,7 +28,7 @@
     "scripts": {
         "dev": "react-scripts start",
         "build": "react-scripts build",
-        "test": "react-scripts test",
+        "test": "react-scripts test --watchAll=false --passWithNoTests",
         "lint": "eslint src --ext .js,.jsx,.ts,.tsx"
     },
     "browserslist": {
