+++ package.json
@@ -10,6 +10,7 @@
         "build": "turbo run build",
         "dev": "turbo run dev --parallel",
         "lint": "turbo run lint",
+        "test": "turbo run test",
         "format": "prettier --write \"**/*.{js,jsx,ts,tsx,md,json}\"",
         "prepare": "husky install",
         "db:start": "docker compose up -d",
