+++ turbo.json
@@ -7,6 +7,9 @@
         "lint": {
             "outputs": []
         },
+        "test": {
+            "outputs": []
+        },
         "dev": {
             "cache": false
         }
